+++ src/Transformable/Primitive/Sphere.hpp
@@ -15,7 +15,7 @@ namespace Transformable {
     namespace Primitive {
         class Sphere : public APrimitive {
             public:
-                Sphere(Point3f, float ray, std::shared_ptr<Material::IMaterial> material);
+                Sphere(Point3f, double ray, std::shared_ptr<Material::IMaterial> material);
                 std::tuple<bool, Display::Color> checkHit(std::unique_ptr<Raytracer::IVector> &) final;
 
             private:
